refactor: update component imports to use common directory and add new Container and Section components for improved layout
